Added basic version of strafer ai, implemented health for players and damage for shots, player and ai can now be shot and recieve damage, ai and bullets despawn when appropriate
Screen boundaries are not yet in place.
Strafer will move in a single direction and will turn towards the player at a slower constant rate, when it is within 0.2 degrees of the player aiming/tracking will be more precise.
Need to implement a warm up timer for ai ships so they don't shoot until after they are fully on screen.